+++ packages/gestalt-codemods/0.111.0-0.112.0/__testfixtures__/text-remove-semibold-dynamic-bold-first.input.js
@@ -1,0 +1,15 @@
+// @flow
+import React from 'react';
+import { Text } from 'gestalt';
+
+export default function BoldText({
+  isCurrentPage,
+}: {
+  isCurrentPage?: boolean,
+}) {
+  return (
+    <Text weight={isCurrentPage ? "bold" : "semibold"} size={isCurrentPage ? "xl" : "md"}>
+      Bold
+    </Text>
+  );
+}