+++ packages/gestalt-codemods/0.111.0-0.112.0/__tests__/heading-remove-semibold.test.js
@@ -1,0 +1,20 @@
+import { defineTest } from 'jscodeshift/dist/testUtils.js';
+
+jest.mock('../heading-remove-semibold', () => {
+  return Object.assign(require.requireActual('../heading-remove-semibold'), {
+    parser: 'flow',
+  });
+});
+
+describe('heading-remove-semibold', () => {
+  ['heading-remove-semibold-dynamic', 'heading-remove-semibold-static'].forEach(
+    test => {
+      defineTest(
+        __dirname,
+        'heading-remove-semibold',
+        { quote: 'single' },
+        test
+      );
+    }
+  );
+});