+++ packages/gestalt-codemods/0.111.0-0.112.0/__tests__/text-remove-semibold.test.js
@@ -1,0 +1,18 @@
+import { defineTest } from 'jscodeshift/dist/testUtils.js';
+
+jest.mock('../text-remove-semibold', () => {
+  return Object.assign(require.requireActual('../text-remove-semibold'), {
+    parser: 'flow',
+  });
+});
+
+describe('text-remove-semibold', () => {
+  [
+    'text-remove-semibold-dynamic-bold-first',
+    'text-remove-semibold-dynamic-semibold-first',
+    'text-remove-semibold-dynamic-normal',
+    'text-remove-semibold-static',
+  ].forEach(test => {
+    defineTest(__dirname, 'text-remove-semibold', { quote: 'single' }, test);
+  });
+});